Created new steps for exporting and downloading csv file for MI CB1 Extract. Download directories created for circleci, docker and selenium. New steps for checking csv file exists and contains data.
Updated version of postgres to match that on production.
Updated readme to reflect changes necessary to run chromedriver.
Updated step to enter a date range.

Added in volumes to the selenium and end to end test containers so can download from headless chrome.
Override downloads variable for chromedriver as cannot create/access the one created in the containers.
Updated .gitignore so that it doesn't prompt saving of downloaded files.

Removed commented out preferences

Updated comment to show JIRA ticket details

Split up last step into two steps to make it easier to debug if there is a problem.

added in MINIMUM_WAIT_TIME variable which was somehow removed through rebase

Created a downloads folder to see if can get this to work on circleci. Removed the loaddata command from the circleci config as this is now done as part of the create_db.sh.

Created a downloads folder to see if can get this to work on circleci. Removed the loaddata command from the circleci config as this is now done as part of the create_db.sh.

set user to root on seleniumchrome image so can write to downloads folder.